requirements.txt
bibtexparser==1.1.0
cssselect==1.0.3
future==0.17.1
Jinja2==2.10
lxml==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.0
pyparsing==2.3.1
pyquery==1.4.0
python-dateutil==2.7.5
pyyaml>=4.2b1
six==1.12.0
